requirements.txt
torch==2.1.0
torchvision==0.16.0
numpy==1.26.0
scikit-learn==1.3.2
PyYAML==6.0.1
tqdm==4.66.1